+++ pysatMissionPlanning/methods/__init__.py
@@ -3,6 +3,10 @@ pysatMissionPlanning.methods is a module that provides
 the methods to interface with numerous empirical model packages
 """
 
-from pysatMissionPlanning.methods import aacgmv2, apexpy, pyglow, spacecraft
+from pysatMissionPlanning.methods import aacgmv2
+from pysatMissionPlanning.methods import apexpy
+from pysatMissionPlanning.methods import plot
+from pysatMissionPlanning.methods import pyglow
+from pysatMissionPlanning.methods import spacecraft
 
-__all__ = ['aacgmv2', 'apexpy', 'pyglow', 'spacecraft']
+__all__ = ['aacgmv2', 'apexpy', 'plot', 'pyglow', 'spacecraft']
